+++ app/src/main/java/com/mapbox/services/android/navigation/testapp/example/ui/navigation/OfflineRouteFinder.kt
@@ -6,10 +6,10 @@ import android.widget.Toast
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.geojson.Point
 import com.mapbox.mapboxsdk.Mapbox
+import com.mapbox.navigation.base.route.NavigationRoute
 import com.mapbox.services.android.navigation.testapp.NavigationApplication
 import com.mapbox.services.android.navigation.testapp.R
 import com.mapbox.services.android.navigation.v5.navigation.MapboxOfflineRouter
-import com.mapbox.navigation.route.offboard.NavigationRoute
 import com.mapbox.services.android.navigation.v5.navigation.OfflineError
 import com.mapbox.services.android.navigation.v5.navigation.OfflineRoute
 import com.mapbox.services.android.navigation.v5.navigation.OnOfflineRouteFoundCallback
@@ -81,7 +81,7 @@ class OfflineRouteFinder(
     private fun buildOfflineRoute(location: Location, destination: Point): OfflineRoute {
         val origin = Point.fromLngLat(location.longitude, location.latitude)
         val bearing = location.bearing.toDouble()
-        return com.mapbox.navigation.route.offboard.NavigationRoute.builder(NavigationApplication.instance)
+        return NavigationRoute.builder(NavigationApplication.instance)
             .origin(origin, bearing, BEARING_TOLERANCE)
             .destination(destination)
             .accessToken(Mapbox.getAccessToken()!!) // OZ: Ok to use "!!". This code should crash if getAccessToken() returns null
