+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/MapboxOffboardRouter.kt
@@ -48,9 +48,6 @@ class MapboxOffboardRouter(
         })
     }
 
-    override fun getRoute(origin: Point, waypoints: List<Point>, listener: Router.RouteListener) {
-    }
-
     override fun cancel() {
         navigationRoute?.cancelCall()
         navigationRoute = null
